refactor(SuggestionProvider): Remove unused code
The SuggestionProvider is no longer used because the search for Drive was repatriated in it rather than in the cozy-bar
