+++ src/drive/lib/FuzzyPathSearch.spec.js
